edited question/answer to question #33 of text
See lines 1128 through 1136.

Exercise 33 had multiple issues. The given function does not actually
pass through the indicated point of the problem statement, and the
tangent/normal lines given in the answer were thus incorrect. The point
was adjusted to lie on the graph of the function, and correct
tangent/normal line equations are given.